tools/google-bert-chinese/requirements.txt
pip
numpy~=1.25

onnx~=1.16
# onnx-weekly
onnxruntime-training==1.18.0
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ORT/pypi/simple/ onnxruntime-training-cpu==1.18.0
torch~=2.3 
torch-ort~=1.17
onnxruntime-training-cpu==1.18.0

transformers[torch]
onnx-simplifier

# install MagicONNX, at model-make.sh

